refactor: move `USDZExporter` into TypeScript
